+++ src/common_clj/time/parser/core.clj
@@ -1,7 +1,9 @@
 (ns common-clj.time.parser.core
   (:require [schema.core :as s])
   (:import (java.time LocalDate)
-           (java.time.format DateTimeFormatter)))
+           (java.time.format DateTimeFormatter)
+           (java.util Date)
+           (java.text SimpleDateFormat)))
 
 (s/defn str->local-date :- LocalDate
   [in-format :- s/Str
@@ -16,3 +18,8 @@
    date :- LocalDate]
   (let [formatter (DateTimeFormatter/ofPattern out-format)]
     (.format date formatter)))
+
+(s/defn date->wire :- s/Str
+  "Convert Date object to ISO-8601 string"
+  [date :- Date]
+  (.format (SimpleDateFormat. "yyyy-MM-dd'T'HH:mm:ssZ") date))